requirements.txt
torch
transformers
Pillow
Flask
